+++ src/main/java/com/example/demo/configuration/filter/CustomAuthenticationFilter.java
@@ -3,8 +3,8 @@ package com.example.demo.configuration.filter;
 import com.example.demo.enums.ContentType;
 import com.example.demo.enums.RequestParameter;
 import com.example.demo.exception.BadPasswordOrEmailException;
-import com.example.demo.model.UserDetailsImpl;
-import com.example.demo.service.TokensService;
+import com.example.demo.model.TeacherUserDetails;
+import com.example.demo.controller.security.TokenProvider;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.http.HttpServletRequest;
@@ -20,7 +20,7 @@ import java.io.IOException;
 @RequiredArgsConstructor
 public class CustomAuthenticationFilter extends UsernamePasswordAuthenticationFilter {
    private final AuthenticationManager authenticationManager;
-   private final TokensService tokensService;
+   private final TokenProvider tokenProvider;
 
     @Override
     public Authentication attemptAuthentication(HttpServletRequest request, HttpServletResponse response) {
@@ -42,8 +42,8 @@ public class CustomAuthenticationFilter extends UsernamePasswordAuthenticationFi
     @Override
     protected void successfulAuthentication(HttpServletRequest request, HttpServletResponse response,
                                             FilterChain chain, Authentication authResult) throws IOException{
-        var user = (UserDetailsImpl) authResult.getPrincipal();
-        var tokens = tokensService.generateTokens(request, user);
+        var user = (TeacherUserDetails) authResult.getPrincipal();
+        var tokens = tokenProvider.generateTokens(request, user);
         response.setContentType(ContentType.JSON.getContentType());
         new ObjectMapper().writeValue(response.getOutputStream(), tokens);
     }
