+++ src/main/java/com/example/demo/configuration/filter/CustomAuthorizationFilter.java
@@ -1,7 +1,7 @@
 package com.example.demo.configuration.filter;
 
 import com.example.demo.enums.AuthorizationType;
-import com.example.demo.service.TokensService;
+import com.example.demo.controller.security.TokenProvider;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
@@ -14,11 +14,11 @@ import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 @RequiredArgsConstructor
 public class CustomAuthorizationFilter extends OncePerRequestFilter {
 
-    private final TokensService tokensService;
+    private final TokenProvider tokenProvider;
     @Override
     protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
         var authorizationHeader = request.getHeader(AUTHORIZATION);
-        if (authorizationHeader != null && authorizationHeader.startsWith(AuthorizationType.BEARER.getPrefix())) tokensService.verifyTokens(authorizationHeader);
+        if (authorizationHeader != null && authorizationHeader.startsWith(AuthorizationType.BEARER.getPrefix())) tokenProvider.verifyTokens(authorizationHeader);
         filterChain.doFilter(request, response);
     }
